arch: arm: function to (re)initialize MPU region for kernel ram
This commit introduces a function to reset the MPU region for
the whole kernel ram to its initial configuration. In addition,
it disables any MPU regions in kernel ram space. The function is
to be used upon context switch, for MPU implementations that
require non-overlapping regions. It is needed only if the kernel
has features (e.g. Thread Stacks or Stack Guards) that generate
MPU regions (partitions) inside the kernel SRAM area. For
consistency, the access permissions of the kernel ram area are
set, also, during initialization.

Signed-off-by: Ioannis Glaropoulos <Ioannis.Glaropoulos@nordicsemi.no>