arch: arm: extend use of arch.mode variable in FP sharing mode
Only a single bit of the 4-byte internal (kernel.arch) state
variable 'mode' is currently used (when we build with
CONFIG_USERSPACE). In this commit we extend the usage of mode
variable, adding an additional bit-flag to track the state
of the floating point context in a particular thread, i.e. to
track whether FP is active or not.

Signed-off-by: Ioannis Glaropoulos <Ioannis.Glaropoulos@nordicsemi.no>