arch: arm: introduce macro to describe the privilege stack alignment
We introduce a new define to describe the alignment of a
privilege stack buffer. This macro is used by the privilege
stack generation script, to determine the required alignment
of threads' privilege stacks when building with support for
user mode.

We cannot use THREAD_MIN_STACK_ALIGN in this case, because the
the privilege stacks do not need to respect the minimum MPU
region alignment requirement, unless, of course, this is
enforced via the MPU Stack Guard feature).

Signed-off-by: Ioannis Glaropoulos <Ioannis.Glaropoulos@nordicsemi.no>