arch: arm: move privilege stack guard right below the priv stack
We are able to use the area rigth before a thread's privileged
stack as guard area to detect stack overflows (when building
with CONFIG_MPU_STACK_GUARD=y). This is a very important
optimization, as we do not need to eat stack memory from each
user thread, and reserve it for a stack-guard.

Signed-off-by: Ioannis Glaropoulos <Ioannis.Glaropoulos@nordicsemi.no>