arch: arm: userspace: program built-in guard in SVC
For ARMv8-M builds supporting stack overflow
detection with the built-in SPLIM registers
(CONFIG_BUILTIN_STACK_GUARD=y), we need to
program the PSPIM during system call setup
in the SVC handler, to point to the start of
the default stack. This is needed to be able
to detect stack overflows in z_arm_do_syscall,
but before the thread switches to using the
privilege stack.

Signed-off-by: Ioannis Glaropoulos <Ioannis.Glaropoulos@nordicsemi.no>