arch: arm: userspace: guard user stack in z_arm_userspace_enter()
In z_arm_userspace_enter(), we set PSPLIM to temporarily guard
the user stack area, after switching PSP back to user stack.
This adds to the safety of the function, since, at that point,
z_arm_userspace_enter() still executes in privileged mode. As
soon as the thread drops privileges, this guarding is not
required any more (following context-switch will remove it).

Signed-off-by: Ioannis Glaropoulos <Ioannis.Glaropoulos@nordicsemi.no>