arch: arm: userspace: no PSPLIM clearing in z_arm_do_syscall() enter
In this commit we remove the PSPLIM clearing when entering
z_arm_do_syscall(). PSPLIM clearing is not strictly required,
and may, even, be overwritten by a swap that will configure
the PSPLIM register to guard the right thread stack (privileged
or user).

We add a comment to stress that PSPLIM still guards the right
thread stack at z_arm_do_syscall() enter, and stress that the
warning on not-using the default stack applies only to the
case where MPU-based stack protection is enabled.

Signed-off-by: Ioannis Glaropoulos <Ioannis.Glaropoulos@nordicsemi.no>