arch: arm: userspace: temporarily guard stack in syscall returning
When returning from a system call, the thread switches
back to using its default stack. Until privileges are
dropped to user mode, we want to guard the default
stack for possible stack overflows. The guard will
remain after returning to user mode, but it won't be
required (user sandboxing is sufficient). The guard
will be removed in the following context-switch.

Signed-off-by: Ioannis Glaropoulos <Ioannis.Glaropoulos@nordicsemi.no>